test: Unify GUI tests
